% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animals.R
\docType{data}
\name{animals}
\alias{animals}
\title{Perfusion measurements from 20 dogs.}
\format{A data table with 11200 observations of 5 variables
\describe{
  \item{animal}{Which animal}
  \item{timestep}{Frame at which the measurement was taken}
  \item{roi}{ROI of the measurement}
  \item{side}{Side of the measurement}
  \item{value}{The actual measurement values}
}}
\usage{
animals
}
\description{
A dataset containing perfusion measurements from 20 dogs.
}
\keyword{datasets}
