% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_params.R
\name{get_perfusion_params}
\alias{get_perfusion_params}
\title{Extract perfusion parameters}
\usage{
get_perfusion_params(flow_scaled_residue, delta_t, density)
}
\arguments{
\item{flow_scaled_residue}{Flow scaled residue function}

\item{delta_t}{Time between Frames}

\item{density}{Density of brain tissue}
}
\value{
List with perfusion parameters
}
\description{
Low level function used to extract perfusion parameters from a flow-scaled residue function
}
