% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolv_reg.R
\name{perfresidue}
\alias{perfresidue}
\title{Extract a flow scaled residue function from AIF and ROI measurements}
\usage{
perfresidue(aif, roi, arrival_frame, delta_t, echo_time, alpha = 0.1,
  lambda = NULL, iterations = 10)
}
\arguments{
\item{aif}{Numeric vector, the arterial input function}

\item{roi}{Numeric vector, the ROI measurement}

\item{arrival_frame}{Integer, the frame number at which the bolus arrived}

\item{delta_t}{Float, the time between two consecutive images}

\item{echo_time}{Float, MRI echo time of the sequence}

\item{alpha}{Float, l1 hyperparameter (l1 vs l2 weight), optional, default 0.1}

\item{lambda}{Float, l1-l2 glmnet hyperparameter. Will be determined via cross-validation
if not provided (recommend)}

\item{iterations}{Integer, the number of iterations used for cross-validation (optional, default 10)}
}
\value{
List with the flow-scaled residue function, the fitted lambda, the regression matrix 
  as well as the signal used (rescaled ROI information)
}
\description{
\code{perfresidue} is a low-level function which givem AIF and ROI measurements 
obtains an estimate for the flow-scaled residue function
}
\details{
We refer to the Review Article by Fieselmann et. al. (see \url{https://doi.org/10.1155/2011/467563})
for detailed definitions. This function implements an improved version of the algorithm described
in the previously mentioned paper, employing machine learning techniques to obtain
better estimates for the flow-scaled residue function.
}
