% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{vetperf.data}
\alias{vetperf.data}
\title{Calculate perfusion parameters}
\usage{
vetperf.data(data, arrival_frame, delta_t, echo_time,
  animalcol = "animal", sidecol = "side", roicol = "roi",
  aifroi = "aif", valcol = "value", baseroi = NULL, density = 1.04,
  lambda = NULL, iterations = 10)
}
\arguments{
\item{data}{Data.table object containing aif and roi measurements}

\item{arrival_frame}{Integer, the frame number at which the bolus arrived}

\item{delta_t}{Float, the time between two consecutive images}

\item{echo_time}{Float, MRI echo time of the sequence}

\item{animalcol}{String, the name of the column holding the animal information}

\item{sidecol}{String, the name of the column indicating which side is measured}

\item{roicol}{String, the name of the column which indicates the ROI}

\item{aifroi}{String, the ROI which contains the AIF measurement (default "aif")}

\item{valcol}{String, the name of the column which holds the data}

\item{baseroi}{String, the name of the white matter roi which should be used for normalization 
(if normalized values should be returned)}

\item{density}{Float, density of the brain (optional)}

\item{lambda}{Float, l1-l2 glmnet hyperparameter. Will be determined via cross-validation
if not provided (recommend)}

\item{iterations}{Integer, the number of iterations used for cross-validation (optional, default 10)}
}
\value{
Data.table object holding perfusion parameters (MTT, CBF, CBV)
}
\description{
\code{vetperf.data} takes a data.table object containing AIF and ROIs measurements
and replaces them with perfusion parameters
}
\details{
High level function to obtain perfusion parameters from aif and roi measurements.
Given a data.table object holding measurements for one or more animals, one or more sides
and one or more ROIs, this functions extracts the associated perfusion parameters 
and optionally normalizes them with respect to a ROI of choice.
The data.table is assumed to have five columns:
  animal: factor indicating the animal
  side: factor indicating the side
  roi: factor indicating the ROI, must include a level corresponding to
       the arterial input function
  timestep: column indicating the frame at which the measurement was taken
  value: column holding the measurements
}
\examples{
perfdata = vetperf.data(animals, arrival_frame = 9, delta_t = 1.6, echo_time = 0.03, baseroi = "cs", iterations = 1)

}
