% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{vetperf.num}
\alias{vetperf.num}
\title{Calculate perfusion parameters from numeric vectors}
\usage{
vetperf.num(aifnum, roinum, arrival_frame, delta_t, echo_time,
  density = 1.04, lambda = NULL, iterations = 10)
}
\arguments{
\item{aifnum}{Numeric vector, the arterial input function}

\item{roinum}{Numeric vector, the ROI measurement}

\item{arrival_frame}{Integer, the frame number at which the bolus arrived}

\item{delta_t}{Float, the time between two consecutive images}

\item{echo_time}{Float, MRI echo time of the sequence}

\item{density}{Float, density of the brain (optional)}

\item{lambda}{Float, l1-l2 glmnet hyperparameter. Will be determined via cross-validation
if not provided (recommend)}

\item{iterations}{Integer, the number of iterations used for cross-validation (optional, default 10)}
}
\value{
List with MTT, CBF, CBV parameters
}
\description{
\code{vetperf.num} takes an arterial input function as well as a ROI measurement and
computes the perfusion parameters MTT, CBF and CBV
}
\details{
Internally, it performs a l1-l2 weighted regression employing cross-validation to obtain the best
fit for the response
}
